Limit the scope of the API: no byte operations or Test methods exposed.
After reading http://blog.golang.org/organizing-go-code I have decided to not
expose the wonderful BytesToUint64, Uint64ToBytes, FmtBytes methods.

We are not a binary-handling library, and they are easy enough to come up with
if they are needed.

Additionally, the TestRecords() method has been made checkRecords() and
unexposed. checkRecords() is used when serializing and deserializing a
message, and on SetRecords, which makes it redundant to give direct
access to it.