Big code refactor. Proper support for Text and URI types.
While looking at the specs I have realized that URI and Text have their
own documents. In fact, text is not just a plain UTF-8 string as I thought,
but something way more complex. This has required implementation of
modules under types/wkt/uri and types/wkt/text

Additionally, while checking out the SmartPoster type specs, I have realized
that a NDEF Message is not the result of parsing a single record (or record chunks),
but that it can have instead several records, with different types each.

This has promted a refactoring to reflect this reality. Much of the funcionality of
message.go has been moved to record.go, and much of the Record has been moved to
record_chunk.go.

There is still work to do (support SmartPoster) and make use of the RecordType interface
to deal with record payloads properly.